Move string interning to util
Code that handles string interning is rather an util functionality than
a core functionality.
