+++ src/cargo/core/compiler/build_context/mod.rs
@@ -2,9 +2,10 @@ use crate::core::compiler::unit_graph::UnitGraph;
 use crate::core::compiler::{BuildConfig, CompileKind, Unit};
 use crate::core::profiles::Profiles;
 use crate::core::PackageSet;
-use crate::core::{InternedString, Workspace};
+use crate::core::Workspace;
 use crate::util::config::Config;
 use crate::util::errors::CargoResult;
+use crate::util::interning::InternedString;
 use crate::util::Rustc;
 use std::collections::HashMap;
 use std::path::PathBuf;
