+++ src/cargo/core/compiler/unit_dependencies.rs
@@ -22,8 +22,9 @@ use crate::core::dependency::DepKind;
 use crate::core::profiles::{Profile, Profiles, UnitFor};
 use crate::core::resolver::features::{FeaturesFor, ResolvedFeatures};
 use crate::core::resolver::Resolve;
-use crate::core::{InternedString, Package, PackageId, PackageSet, Target, Workspace};
+use crate::core::{Package, PackageId, PackageSet, Target, Workspace};
 use crate::ops::resolve_all_features;
+use crate::util::interning::InternedString;
 use crate::util::Config;
 use crate::CargoResult;
 use log::trace;
