+++ src/cargo/core/resolver/features.rs
@@ -42,7 +42,8 @@ use crate::core::compiler::{CompileKind, RustcTargetData};
 use crate::core::dependency::{DepKind, Dependency};
 use crate::core::resolver::types::FeaturesSet;
 use crate::core::resolver::{Resolve, ResolveBehavior};
-use crate::core::{FeatureValue, InternedString, PackageId, PackageIdSpec, PackageSet, Workspace};
+use crate::core::{FeatureValue, PackageId, PackageIdSpec, PackageSet, Workspace};
+use crate::util::interning::InternedString;
 use crate::util::CargoResult;
 use std::collections::{BTreeSet, HashMap, HashSet};
 use std::rc::Rc;
