+++ src/cargo/core/workspace.rs
@@ -13,11 +13,12 @@ use crate::core::features::Features;
 use crate::core::registry::PackageRegistry;
 use crate::core::resolver::features::RequestedFeatures;
 use crate::core::resolver::ResolveBehavior;
-use crate::core::{Dependency, InternedString, PackageId, PackageIdSpec};
+use crate::core::{Dependency, PackageId, PackageIdSpec};
 use crate::core::{EitherManifest, Package, SourceId, VirtualManifest};
 use crate::ops;
 use crate::sources::PathSource;
 use crate::util::errors::{CargoResult, CargoResultExt, ManifestError};
+use crate::util::interning::InternedString;
 use crate::util::paths;
 use crate::util::toml::{read_manifest, TomlProfiles};
 use crate::util::{Config, Filesystem};
