+++ src/cargo/util/command_prelude.rs
@@ -1,9 +1,9 @@
 use crate::core::compiler::{BuildConfig, MessageFormat};
-use crate::core::InternedString;
 use crate::core::Workspace;
 use crate::ops::{CompileFilter, CompileOptions, NewOptions, Packages, VersionControl};
 use crate::sources::CRATES_IO_REGISTRY;
 use crate::util::important_paths::find_root_manifest_for_wd;
+use crate::util::interning::InternedString;
 use crate::util::{paths, toml::TomlProfile, validate_package_name};
 use crate::util::{
     print_available_benches, print_available_binaries, print_available_examples,
