+++ src/cargo/util/interning.rs
